rector: location organize package
