+++ src/main/java/com/rgv04/hr/domain/departament/Departament.java
@@ -7,7 +7,7 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
-import com.rgv04.hr.domain.location.Location;
+import com.rgv04.hr.domain.model.Location;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
