+++ src/main/java/com/rgv04/hr/domain/departament/DepartamentModel.java
@@ -3,7 +3,7 @@ package com.rgv04.hr.domain.departament;
 import org.springframework.hateoas.RepresentationModel;
 import org.springframework.hateoas.server.core.Relation;
 
-import com.rgv04.hr.domain.location.LocationModel;
+import com.rgv04.hr.domain.dto.LocationModel;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
