--- src/main/java/com/rgv04/hr/domain/location/LocationRepositoryQueries.java
@@ -1,9 +1,0 @@
-package com.rgv04.hr.domain.location;
-
-import java.util.List;
-
-public interface LocationRepositoryQueries {
-
-	List<Location> ListByFilter(LocationFilter filter);
-
-}
