+++ src/main/java/com/rgv04/hr/domain/service/LocationService.java
@@ -1,10 +1,13 @@
-package com.rgv04.hr.domain.location;
+package com.rgv04.hr.domain.service;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.rgv04.hr.domain.dto.LocationFilter;
+import com.rgv04.hr.domain.model.Location;
+import com.rgv04.hr.domain.repository.LocationRepository;
 import com.rgv04.hr.exception.BusinessException;
 
 @Service
